make github happy add cgo support
